Update modprobe config file status in README
The modprobe configuration file that ships with the driver is designed to run the device pairing utility whenever the ipheth module is loaded by the kernel. As this is already done (arguably in a better way as it happens per device) by the udev rules that ship with ipheth-pair, installing this file is no longer required and actually not really recommended.